修复: 系列编译警告（Qt6 弃用 API 与 MSVC C4xxx 警告）
任务类型: 修复
内容:
- Qt6 弃用 API: setFontFamily→setFontFamilies（Qt5.14+ 可用，无需版本分支）;
  setNamedColor→DA::compat::setNamedColor; canConvert(QMetaType::xx)→canConvert<T>()
- DAPybind11QtCaster.hpp: 移除 sscanf，改用可移植十六进制解析器（C4996）
- 修复 qHash 系列函数 size_t→uint 隐式截断（C4267），含排查中发现的
  DAPyNodeConnection.cpp / DAPyNodeMetaData.cpp 同类隐患
- 修复 DAChartUtil.h 模板与 DAPyDataFrame.cpp 有符号/无符号比较（C4018）
- 移除 catch 未引用变量（C4101）与 DADialogChartGuide 空 switch 占位（C4065）
- DAGraphicsResizeableItem XML 序列化改为直接访问 mControlPointSize，
  不再调用自身已弃用的 getControlerSize/setControlerSize
- ads::CDockWidget 改用带 manager 的新构造函数（旧构造函数已被 ADS 弃用）

验证: Qt6.7.3 + VS2019 Release 全量构建通过，全部警告清零；
Qt5.14.2 兼容性经 Qt5 头文件核验（setFontFamilies/fontFamilies 自 5.14 可用），
因本机 Qt5 第三方库安装不完整（缺 DAWidgets）未做 Qt5 实机构建

涉及文件: DAUtils(DAMarkdownHighlighter/DAXMLFileInterface),
DAAxOfficeWrapper(DAAxObjectExcelSheetWrapper), DAPyBindQt(DAPybind11QtCaster.hpp/
DAPyObjectWrapper/DAPyDataFrame), DAPyScripts(DAPyScriptsDataFrame),
DAPyWorkFlow(DAPyLinkPoint/DAPyNode/DAPyNodeConnection/DAPyNodeMetaData/
DAPyWorkFlowPythonBinding), DAGraphicsView(DAGraphicsStandardTextItem/
DAGraphicsResizeableItem), DAFigure(DAChartUtil.h), DAGui(DATableCellStyle/
DADataTableModel/DAPyDataFrameTableModel/DADialogChartGuide),
DAInterface(DADockingAreaInterface), APP(DAAppDockingArea)

+++ src/DAPyBindQt/DAPyObjectWrapper.cpp
@@ -486,7 +486,7 @@ bool DAPyObjectWrapper::operator<(const DAPyObjectWrapper& obj) const
  */
 uint qHash(const DAPyObjectWrapper& obj, uint seed)
 {
-    return ::qHash(reinterpret_cast<quintptr>(obj.object().ptr()), seed);
+    return static_cast< uint >(::qHash(reinterpret_cast<quintptr>(obj.object().ptr()), seed));
 }
 
 /**
